+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerDefaultDateTime.jsx
@@ -19,6 +19,7 @@ const DateTimePickerExampleDefaultDateTime = () => {
       <Field label="Enter Date/Time" htmlFor="defaultDateTime">
         <DateTimePicker
           name="date-time-picker-example"
+          ariaLabel='Enter Date/Time'
           dateInputAttributes={{ id: 'defaultDateTime' }}
           onChange={handleDateTimeChange}
           value={moment().format()}