+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerIncomplete.jsx
@@ -17,6 +17,7 @@ const DateTimePickerExampleIncomplete = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="default-incomplete">
         <DateTimePicker
+          ariaLabel='Enter Date/Time'
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'default-incomplete' }}
           onChange={handleDateTimeChange}