+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerInvalid.jsx
@@ -22,6 +22,7 @@ const DateTimePickerExampleInvalid = () => {
       <button type="button" onClick={handleInvalidButtonClick}>Toggle Validity</button>
       <Field label="Enter Date/Time" htmlFor="default-invalid" error="Error Message" isInvalid={isInvalid}>
         <DateTimePicker
+          ariaLabel='Enter Date/Time'
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'default-invalid' }}
           onChange={handleDateTimeChange}