+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerMinMax.jsx
@@ -18,6 +18,7 @@ const DateTimePickerExampleMinMax = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="min-max-dates">
         <DateTimePicker
+          ariaLabel='Enter Date/Time'
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'min-max-dates' }}
           onChange={handleDateTimeChange}