+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Changed
+  * Added fix for field label SR announce.
+  
+* Changed
   * Changed `up`,`down` key functionality to same as `+`,`-` functionality to make uniform.
   * Changed `terra-time-input` type `number` instead of `text`.